Bugs are fixed again for tasks 0 and 1

+++ 0x15-file_io/0-read_textfile.c
@@ -5,7 +5,6 @@
  * POSIX standard output
  * @filename: file name
  * @letters: the number of letters
- *
  * Return: 1 for success or -1 for failure
  */
 ssize_t read_textfile(const char *filename, size_t letters)
@@ -41,5 +40,5 @@ ssize_t read_textfile(const char *filename, size_t letters)
 
 	if (fwr <= 0)
 		return (0);
-	return ((ssize_t)fwr);
+	return (fwr);
 }
